+++ .eslintignore
@@ -17,10 +17,7 @@ wagtail/admin/templates/wagtailadmin/edit_handlers/inline_panel.js
 wagtail/contrib/search_promotions/templates/wagtailsearchpromotions/includes/searchpromotions_formset.js
 wagtail/users/templates/wagtailusers/groups/includes/page_permissions_formset.js
 wagtail/snippets/templates/wagtailsnippets/chooser/chosen.js
-wagtail/images/templates/wagtailimages/chooser/image_chosen.js
-wagtail/images/templates/wagtailimages/chooser/chooser.js
 wagtail/search/templates/wagtailsearch/queries/chooser/chooser.js
-wagtail/images/templates/wagtailimages/chooser/select_format.js
 wagtail/embeds/templates/wagtailembeds/chooser/embed_chosen.js
 wagtail/embeds/templates/wagtailembeds/chooser/chooser.js
 wagtail/documents/templates/wagtaildocs/chooser/chooser.js
