+++ wagtail/images/static_src/wagtailimages/js/hallo-plugins/hallo-wagtailimage.js
@@ -26,6 +26,7 @@
                     insertionPoint = $(lastSelection.endContainer).parentsUntil('[data-hallo-editor]').last();
                     return ModalWorkflow({
                         url: window.chooserUrls.imageChooser + '?select_format=true',
+                        onload: IMAGE_CHOOSER_MODAL_ONLOAD_HANDLERS,
                         responses: {
                             imageChosen: function(imageData) {
                                 var elem;
