+++ wagtail/images/static_src/wagtailimages/js/image-chooser.js
@@ -7,6 +7,7 @@ function createImageChooser(id) {
     $('.action-choose', chooserElement).on('click', function() {
         ModalWorkflow({
             url: window.chooserUrls.imageChooser,
+            onload: IMAGE_CHOOSER_MODAL_ONLOAD_HANDLERS,
             responses: {
                 imageChosen: function(imageData) {
                     input.val(imageData.id);
