--- wagtail/images/templates/wagtailimages/chooser/image_chosen.js
@@ -1,4 +1,0 @@
-function(modal, jsonData) {
-    modal.respond('imageChosen', jsonData['result']);
-    modal.close();
-}
