--- wagtail/images/templates/wagtailimages/chooser/select_format.js
@@ -1,11 +1,0 @@
-function(modal) {
-    $('form', modal.body).on('submit', function() {
-        var formdata = new FormData(this);
-
-        $.post(this.action, $(this).serialize(), function(response){
-            modal.loadResponseText(response);
-        }, 'text');
-
-        return false;
-    });
-}
